Transferred error handling logic for posts over to users

+++ userRoute/router.js
@@ -8,7 +8,7 @@ function capitalName(req, res, next) {
   if (name === name.toUpperCase()) {
     next();
   } else {
-    req.body.name = req.body.name.toUpperCase();
+    req.body.name = name.toUpperCase();
     next();
   }
 }
@@ -18,7 +18,7 @@ router.get("/", async (req, res) => {
     const users = await db.get();
     res.status(200).json(users);
   } catch (error) {
-    res.status(500).json({ message: "Error" });
+    res.status(500).json({ message: "The user information could not be retrieved." });
   }
 });
 
@@ -26,9 +26,15 @@ router.get("/:id", async (req, res) => {
   const { id } = req.params;
   try {
     const user = await db.getById(id);
-    res.status(201).json(user);
+    if (user) {
+      res.status(200).json(user);
+    } else {
+      res
+        .status(404)
+        .json({ message: "The user with the specified ID does not exist." });
+    }
   } catch (error) {
-    res.status(500).json({ message: "error" });
+    res.status(500).json({ message: "The user's information could not be retrieved." });
   }
 });
 
@@ -46,9 +52,15 @@ router.post("/", capitalName, async (req, res) => {
   const user = req.body;
   try {
     const newUser = await db.insert(user);
-    res.status(200).json(newUser);
+    if (newUser) {
+      res.status(201).json(newUser);
+    } else {
+      res.status(500).json({
+        message: "There was an error while saving the post to the database"
+      });
+    }
   } catch (error) {
-    res.status(500).json({ message: "Error" });
+    res.status(500).json({ message: "Something went wrong when you made your request" });
   }
 });
 
@@ -56,22 +68,51 @@ router.delete("/:id", async (req, res) => {
   const { id } = req.params;
   try {
     const user = await db.getById(id);
-    const deleted = await db.remove(id);
-    res.status(200).json(user);
+    if (user) {
+      const deleted = await db.remove(id);
+      if (deleted) {
+        res.status(200).json(user);
+      } else {
+        res.status(500).json({ error: "The post could not be removed" });
+      }
+    } else {
+      res
+    .status(404)
+    .json({ message: "The user with the specified ID does not exist." });
+    }
+    
   } catch (error) {
-    res.status(500).json({ message: "Error" });
+    res.status(500).json({ message: "Something went wrong when you made your request." });
   }
 });
 
 router.put("/:id", capitalName, async (req, res) => {
   const user = req.body;
   const { id } = req.params;
-  try {
-    const edited = await db.update(id, user);
-    const newUser = await db.getById(id);
-    res.status(202).json(newUser);
-  } catch (error) {
-    res.status(500).json({ message: "Error" });
+  if (!user.name) {
+    res
+      .status(400)
+      .json({ message: "Please provide a name for this user" });
+  } else {
+    try {
+      const newUser = await db.getById(id);
+      if (newUser) {
+        const edited = await db.update(id, user);
+        if (edited) {
+          res.status(200).json(newUser);
+        } else {
+          res.status(500).json({
+            message: "The post information could not be modified."
+          });
+        }
+      } else {
+        res
+        .status(404)
+        .json({ message: "The post with the specified ID does not exist." });
+      }
+    } catch (error) {
+      res.status(500).json({ message: "Something went wrong when you made your request." });
+    }
   }
 });
